Allow clock systems for domain offsets
If the domain offsets are ordered in a clock system instead of
starting with globalDomain.getOffset() (could be at (0,0,0) )
at file (0,0,0), this patch allows to read them.

First, the file with the lowest offset (aka
globalDomain.getOffset() ) has to be found. Afterwards,
the same divide&conquer algorithm looks for the top-left corner,
but only in the "upper" or "lower" volume, seen from the "first"
file.

At the end, the real read algorithm will walk over an "optimal"
number of files for requested planes. If no plane was requested
but a general volume, the algorithm (should) still work but is
less efficient if one is not looking for the full domain
(same behaviour as before).