ColTypeString::getSize() as strlen()
make getSize for fixed length strings to behave like
strlen() so setting and getting the length of a c-string
is defined in the same way.

for variable length string, we forward the behaviour of
HDF5, which returns the size of a char pointer for all
lengths. also, we need to use the API H5Tis_variable_str
to check the type correctly (the previous check was
wrong).