Remember HDF5 Version
This commit adds support to libSplash to remember
the HDF5 version it was build against.

This is interesting, since depending software usually
does not include `<hdf5.h>` and can not query the
version of HDF5 directly via the includes.

That does not prevent users to compile HDF5 and libSplash
both with shared libs and to mix various versions, but
the `SPLASH_HDF_VERSION` will still point to the version
Splash was originially build (and should be used) with.