Close #173: Parallel Write of Global Attributes
When writing global attributes in parallel mode,
one would expect them to be under the root `/` path
(and to be identical for each rank).

This closes #173 and is a requirement to allow
implementations to fulfill the openPMD standard.

Serial writes are not affected and still write to
`/custom` (each of the ranks). A later refactoring
could consider just to let one rank write to `/`
here, too (and to ingore the others assuming an
actual "global" == identical value).