Fix Public Wshadow Warnings
This commit fixes warnings in public header files
caused by libSplash in external projects (as in PIConGPU)
when compiled with `-Wshadow`.

Since the affected classes are used as structs (purely public
members) refining those in a `member_m` naming scheme would
be more intrusive then changing the naming in their
contructors.

Generally it is pretty annoying that `-Wshadow` also calls
out contructors but keeping those clean allows to find more
serious bugs with it.

Related to #190 but does not yet fix the Splash compile
itself (non-public headers and implementations).