Werror: On Travis Only
The `-Werror` flag should not be shipped to users.
We will enforce it manually for travis-ci testing so the
(limited) set of compilers we test stays warning-free.

Nevertheless, a (new) warning on a system or compiler we did not test
is not necessarily fatal and would be a showstopper for new users.